+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Expose Convert and InterpretAs MAML nodes [#125](https://github.com/geotrellis/maml/pull/125)
 
 ## [0.6.1] - 2020-06-23
 ### Added